1.限制主线程队列的执行时间保证流畅 2.地形计算改为非递归方式 3.修复了内存溢出的bug(UnloadUnusedAssets()) 4.修改了各模块伪多线程的交互方式(改为单任务队列保证任务先后顺序)

+++ Assets/Scripts/SE/TerrainImpacts/Basic/BasicRandomAdjust.cs
@@ -15,11 +15,11 @@ namespace SE.TerrainImpacts {
                 RandomRange = (ulong)System.Math.Max(Data.Region.x2 - Data.Region.x1, Data.Region.y2 - Data.Region.y1) / 5,
                 t = RandomRange / 2;
 
-            basemap[1] += (long)Data.Seed[1].NextRandomNum(RandomRange) - (long)t;
-            basemap[3] += (long)Data.Seed[2].NextRandomNum(RandomRange) - (long)t;
-            basemap[4] += (long)Data.Seed[0].NextRandomNum(RandomRange) - (long)t;
-            basemap[5] += (long)Data.Seed[3].NextRandomNum(RandomRange) - (long)t;
-            basemap[7] += (long)Data.Seed[4].NextRandomNum(RandomRange) - (long)t;
+            basemap[1] += (long)(Data.Seed[1].NextRandomNum(RandomRange) - t);
+            basemap[3] += (long)(Data.Seed[2].NextRandomNum(RandomRange) - t);
+            basemap[4] += (long)(Data.Seed[0].NextRandomNum(RandomRange) - t);
+            basemap[5] += (long)(Data.Seed[3].NextRandomNum(RandomRange) - t);
+            basemap[7] += (long)(Data.Seed[4].NextRandomNum(RandomRange) - t);
         }
 	}
 }
